# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Método Ponto B - RE/MAX SC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /franchises:
    get:
      operationId: listFranchises
      tags: [franchises]
      summary: List all franchises
      responses:
        "200":
          description: List of franchises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Franchise"
    post:
      operationId: createFranchise
      tags: [franchises]
      summary: Create a franchise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FranchiseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Franchise"
  /franchises/{id}:
    get:
      operationId: getFranchise
      tags: [franchises]
      summary: Get a franchise by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Franchise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Franchise"
    patch:
      operationId: updateFranchise
      tags: [franchises]
      summary: Update a franchise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FranchiseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Franchise"
...
components:
  schemas:
...
    Franchise:
      type: object
      required: [id, name, city, state, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        state:
          type: string
        brokerOwnerName:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    FranchiseInput:
      type: object
      required: [name, city, state]
      properties:
        name:
          type: string
        city:
          type: string
        state:
          type: string
        brokerOwnerName:
          type: string
    FranchiseUpdate:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        state:
          type: string
        brokerOwnerName:
          type: string
        active:
          type: boolean
...