# lib/api-zod/src/generated/types/franchiseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Método Ponto B - RE/MAX SC API
 * OpenAPI spec version: 0.1.0
 */

export interface FranchiseInput {
  name: string;
  city: string;
  state: string;
  brokerOwnerName?: string;
}